Zei code for ABAR to BAR conversion note
The ABAR to BAR conversion is implemented very similar to the BAR to
ABAR one. The Zero Knowledge Proof contains the equality proof of both
amount and asset type. The user is expected to present the proof of
presence of ABAR on the ledger by disclosing its ATxoSID and knowledge of
the randomizer by presenting the signature with the diversified private
key.